Refactor: Move repository CRUD from SettingsPage to RegistryPage
- Move repository add/delete functionality from SettingsPage to RegistryPage
- SettingsPage now focuses only on global settings (Cachix, Attic)
- RegistryPage handles all repository management operations
- Updated routing to point repository operations to RegistryPage
- Fixed import issues with field accessors for record types
- Maintains unified repository UI with collapsible add form
