+++ nix/modules/flake-parts/haskell.nix
@@ -134,7 +134,7 @@
                     curl -X POST -k \
                       --data-urlencode "name=${repo.name}" \
                       --data-urlencode "cloneUrl=${repo.value}" \
-                      "https://${host}:${port}/s/repo"
+                      "https://${host}:${port}/r/add"
                   '')
                   (lib.attrsToList defaultRepos) +
                 # The cachix token here is for a dummy cache, managed by Srid.
@@ -142,7 +142,7 @@
                   curl -X POST -k \
                     --data-urlencode "cachixName=scratch-vira-dev" \
                     --data-urlencode "authToken=eyJhbGciOiJIUzI1NiJ9.eyJqdGkiOiI5NDI4ZjhkZi1mZWM5LTQ1ZjctYjMzYi01MTFiZTljNTNkNjciLCJzY29wZXMiOiJjYWNoZSJ9.WgPWUSYIie2rUdfuPqHS5mxrkT0lc7KIN7QPBPH4H-U" \
-                    "https://${host}:${port}/s/cachix"
+                    "https://${host}:${port}/settings/cachix"
                 '';
               meta.description = ''
                 A post-run setup script for Vira, that initialises a few repos and a dummy cachix.
