+++ src/Vira/App/LinkTo/Type.hs
@@ -13,6 +13,8 @@ data LinkTo
   | RepoListing
   | Repo RepoName
   | RepoUpdate RepoName
+  | RepoAdd
+  | RepoDelete
   | Build RepoName BranchName
   | RepoBranch RepoName BranchName
   | Job JobId
@@ -35,13 +37,15 @@ linkShortTitle = \case
   RepoListing -> "Repositories"
   Repo name -> toText . toString $ name
   RepoUpdate _ -> "Update" -- unused
+  RepoAdd -> "Add Repository"
+  RepoDelete -> "Delete Repository"
   Build _ _ -> "Build" -- unused
   RepoBranch _ branchName -> toText . toString $ branchName
   Job jobId -> "Job " <> show jobId
   JobLog jobId -> "Job Log " <> show jobId
   JobLogStream jobId -> "Job Log Stream " <> show jobId
   Kill _ -> "Kill" -- unused
-  Settings -> "Settings"
+  Settings -> "Global Settings"
   SettingsCachix -> "Cachix Settings"
   SettingsDeleteCachix -> "Delete Cachix Settings"
   SettingsAttic -> "Attic Settings"
